(PC-12791)[API]sendinblue email offer rejection pro
